fix : 도메인 정리

+++ ServerCore/Sys_ConnectResponsePacket.h
@@ -5,8 +5,8 @@
 #pragma pack(push,1)
 struct SYS_CONNECT_RESPONSE_PACKET : public PACKET_HEADER
 {
-	UINT32 ClientId = { 0 };
+	uint32 ClientId = { 0 };
 };
 
-const size_t SYS_CONNECT_RESPONSE_PACKET_SIZE = sizeof(SYS_CONNECT_RESPONSE_PACKET);
+const uint32 SYS_CONNECT_RESPONSE_PACKET_SIZE = sizeof(SYS_CONNECT_RESPONSE_PACKET);
 #pragma pack(pop)
